backend/requirements.txt
Flask==2.2.5
flask-jwt-extended==4.4.4
pymongo==4.4.0
python-dotenv==1.0.0
pandas==2.1.2
scikit-learn==1.3.2
joblib==1.2.0
flask-cors==4.0.0
werkzeug==2.2.3
